+++ modules/apps/collaboration/wiki/wiki-uad/src/main/java/com/liferay/wiki/uad/anonymizer/BaseWikiNodeUADAnonymizer.java
@@ -1,0 +1,84 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.wiki.uad.anonymizer;
+
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.model.User;
+
+import com.liferay.user.associated.data.anonymizer.DynamicQueryUADAnonymizer;
+
+import com.liferay.wiki.model.WikiNode;
+import com.liferay.wiki.service.WikiNodeLocalService;
+import com.liferay.wiki.uad.constants.WikiUADConstants;
+
+import org.osgi.service.component.annotations.Reference;
+
+import java.util.Arrays;
+import java.util.List;
+
+/**
+ * Provides the base implementation for the wiki node UAD anonymizer.
+ *
+ * <p>
+ * This implementation exists only as a container for the default methods
+ * generated by ServiceBuilder. All custom service methods should be put in
+ * {@link WikiNodeUADAnonymizer}.
+ * </p>
+ *
+ * @author Brian Wing Shun Chan
+ * @generated
+ */
+public abstract class BaseWikiNodeUADAnonymizer
+	extends DynamicQueryUADAnonymizer<WikiNode> {
+	@Override
+	public void autoAnonymize(WikiNode wikiNode, long userId, User anonymousUser)
+		throws PortalException {
+		if (wikiNode.getUserId() == userId) {
+			wikiNode.setUserId(anonymousUser.getUserId());
+			wikiNode.setUserName(anonymousUser.getFullName());
+		}
+
+		if (wikiNode.getStatusByUserId() == userId) {
+			wikiNode.setStatusByUserId(anonymousUser.getUserId());
+			wikiNode.setStatusByUserName(anonymousUser.getFullName());
+		}
+
+		wikiNodeLocalService.updateWikiNode(wikiNode);
+	}
+
+	@Override
+	public void delete(WikiNode wikiNode) throws PortalException {
+		wikiNodeLocalService.deleteNode(wikiNode);
+	}
+
+	@Override
+	public List<String> getNonanonymizableFieldNames() {
+		return Arrays.asList("name", "description");
+	}
+
+	@Override
+	protected ActionableDynamicQuery doGetActionableDynamicQuery() {
+		return wikiNodeLocalService.getActionableDynamicQuery();
+	}
+
+	@Override
+	protected String[] doGetUserIdFieldNames() {
+		return WikiUADConstants.USER_ID_FIELD_NAMES_WIKI_NODE;
+	}
+
+	@Reference
+	protected WikiNodeLocalService wikiNodeLocalService;
+}
